+++ test/runTests.ts
@@ -2,6 +2,7 @@ import './cellView.test';
 import './clipboard.test';
 import './dataReader.test';
 import './messages.test';
+import './parsing.test';
 import './profile.test';
 import './tableLayout.test';
 import './tableState.test';